video: cfb_console: fix 32-bit display on 64-bit architectures
"unsigned long" is a lousy data type when it comes to match peripheral
hardware registers with a fixed size.
Just do the obvious and match a 32-bit display format with an "u32"
data type for casting.
This fixes the logo display on 64-bit architectures, which produced
a black line on the right side of the logo with non-black backgrounds.

Change-Id: I292c2a256a2ba8714c6d7afdc0e77b92bb6b383f
Signed-off-by: Andre Przywara <andre.przywara@arm.com>